Address the review: drop lists, shipped API, override order, light header
* registerItemsShortcuts_ descends into the drop list of every button, so a
  drop-only item with a default shortcut is bound too;
* setupCommonPluginsShortcuts() shipped in v3.1.3.663, so its header stays with
  a [[deprecated]] empty body instead of being deleted;
* the doc comment of registerItemsShortcuts_ says that it runs after
  setupShortcuts_ and how to change the default of an item;
* ShortcutKey, ShortcutCategory and Shortcut move to the light MRShortcutKey.h,
  so MRRibbonSchema.h does not pull MRShortcutManager.h into every user;
* the two P* expectations of the new tests are skipped on wasm, where
  getGlfwModPrimaryCtrl() / getGlfwKeyDelete() call the page function is_mac()
  that the MRTest harness does not define.
