+++ source/MRCommonPlugins/MRCommonPlugins.vcxproj.filters
@@ -60,6 +60,7 @@
     <ClCompile Include="Voxels\MROpenRawVoxelsPlugin.cpp">
       <Filter>Voxels</Filter>
     </ClCompile>
+    <ClCompile Include="MRCommonPluginsShortcuts.cpp" />
     <ClCompile Include="Basic\MRCameraOrientationPlugin.cpp">
       <Filter>Basic</Filter>
     </ClCompile>
@@ -102,6 +103,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="exports.h" />
+    <ClInclude Include="MRCommonPluginsShortcuts.h" />
     <ClInclude Include="Basic\MRMoveObjectByMouse.h">
       <Filter>Basic</Filter>
     </ClInclude>
