+++ source/MRCommonPlugins/MRCommonPluginsShortcuts.cpp
@@ -1,0 +1,10 @@
+#include "MRCommonPluginsShortcuts.h"
+
+namespace MR
+{
+
+void setupCommonPluginsShortcuts()
+{
+}
+
+} //namespace MR
