+++ source/MRCommonPlugins/MRCommonPluginsShortcuts.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "MRCommonPlugins/exports.h"
+
+namespace MR
+{
+
+/// does nothing: the items bind their default shortcuts from items.json themselves (see MenuItemShortcut)
+[[deprecated( "the items bind their default shortcuts from items.json themselves, remove the call" )]]
+MRCOMMONPLUGINS_API void setupCommonPluginsShortcuts();
+
+} //namespace MR
